refactor: consolidate trust API clients and types (#779)
Move shared trust API functions and types to the @/api layer (api/trust.ts,
api/endorsements.ts) so both the endorsements and verification features can
import them without violating the no-cross-feature-import ESLint rule.

- @/api/endorsements.ts: canonical home for getMyEndorsements, Endorsement,
  EndorsementsListResponse (renamed/moved from verification/api/client.ts)
- @/api/trust.ts: canonical home for budget, invite, and revoke functions/types
- trust/api/client.ts: slimmed to trust-only concerns; re-exports shared items
- trust/api/queries.ts: adds useMyEndorsementsList; useAcceptInvite now
  invalidates ['trust-endorsements'] and ['verification-status']; all keys use
  trust-* naming convention
- endorsements/api/: re-exports from @/api layer; hooks use trust-* query keys
  eliminating the cache-key drift that caused stale-cache bugs
- endorsements/types.ts: re-exports from @/api/trust and @/api/endorsements
- verification/api/queries.ts: imports getMyEndorsements from @/api/endorsements
- Endorse.page.tsx: imports hooks directly from @/features/trust (the page layer
  is allowed to import from features)
