+++ web/src/features/endorsements/api/queries.ts
@@ -1,19 +1,25 @@
+/**
+ * Endorsement query hooks — delegate to shared @/api layer.
+ *
+ * Query keys use the trust naming convention ('trust-*') so all features share
+ * the same cache entries and avoid stale-cache bugs.
+ */
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import type { CryptoModule } from '@/providers/CryptoProvider';
-import type { CreateInvitePayload } from '../types';
+import { getMyEndorsements } from '@/api/endorsements';
 import {
   acceptInvite,
   createInvite,
-  getMyEndorsements,
-  getMyInvites,
-  getTrustBudget,
+  getMyBudget,
+  listMyInvites,
   revokeEndorsement,
-} from './client';
+  type CreateInvitePayload,
+} from '@/api/trust';
+import type { CryptoModule } from '@/providers/CryptoProvider';
 
 export function useTrustBudget(
   deviceKid: string | null,
   privateKey: CryptoKey | null,
-  crypto: CryptoModule | undefined
+  crypto: CryptoModule | null | undefined
 ) {
   return useQuery({
     queryKey: ['trust-budget', deviceKid],
@@ -21,7 +27,7 @@ export function useTrustBudget(
       if (!deviceKid || !privateKey || !crypto) {
         throw new Error('Not authenticated');
       }
-      return getTrustBudget(deviceKid, privateKey, crypto);
+      return getMyBudget(deviceKid, privateKey, crypto);
     },
     enabled: Boolean(deviceKid && privateKey && crypto),
     staleTime: 30_000,
@@ -31,10 +37,10 @@ export function useTrustBudget(
 export function useMyEndorsementsList(
   deviceKid: string | null,
   privateKey: CryptoKey | null,
-  crypto: CryptoModule | undefined
+  crypto: CryptoModule | null | undefined
 ) {
   return useQuery({
-    queryKey: ['my-endorsements', deviceKid],
+    queryKey: ['trust-endorsements', deviceKid],
     queryFn: async () => {
       if (!deviceKid || !privateKey || !crypto) {
         throw new Error('Not authenticated');
@@ -49,15 +55,15 @@ export function useMyEndorsementsList(
 export function useMyInvites(
   deviceKid: string | null,
   privateKey: CryptoKey | null,
-  crypto: CryptoModule | undefined
+  crypto: CryptoModule | null | undefined
 ) {
   return useQuery({
-    queryKey: ['my-invites', deviceKid],
+    queryKey: ['trust-invites', deviceKid],
     queryFn: async () => {
       if (!deviceKid || !privateKey || !crypto) {
         throw new Error('Not authenticated');
       }
-      return getMyInvites(deviceKid, privateKey, crypto);
+      return listMyInvites(deviceKid, privateKey, crypto);
     },
     enabled: Boolean(deviceKid && privateKey && crypto),
     staleTime: 30_000,
@@ -70,7 +76,7 @@ export function useCreateInvite(deviceKid: string, privateKey: CryptoKey, crypto
     mutationFn: (payload: CreateInvitePayload) =>
       createInvite(deviceKid, privateKey, crypto, payload),
     onSuccess: () => {
-      void queryClient.invalidateQueries({ queryKey: ['my-invites'] });
+      void queryClient.invalidateQueries({ queryKey: ['trust-invites'] });
     },
   });
 }
@@ -80,9 +86,10 @@ export function useAcceptInvite(deviceKid: string, privateKey: CryptoKey, crypto
   return useMutation({
     mutationFn: (inviteId: string) => acceptInvite(deviceKid, privateKey, crypto, inviteId),
     onSuccess: () => {
-      void queryClient.invalidateQueries({ queryKey: ['my-endorsements'] });
-      void queryClient.invalidateQueries({ queryKey: ['my-invites'] });
+      void queryClient.invalidateQueries({ queryKey: ['trust-endorsements'] });
+      void queryClient.invalidateQueries({ queryKey: ['trust-invites'] });
       void queryClient.invalidateQueries({ queryKey: ['trust-budget'] });
+      void queryClient.invalidateQueries({ queryKey: ['trust-scores'] });
       void queryClient.invalidateQueries({ queryKey: ['verification-status'] });
     },
   });
@@ -97,8 +104,9 @@ export function useRevokeEndorsement(
   return useMutation({
     mutationFn: (subjectId: string) => revokeEndorsement(deviceKid, privateKey, crypto, subjectId),
     onSuccess: () => {
-      void queryClient.invalidateQueries({ queryKey: ['my-endorsements'] });
+      void queryClient.invalidateQueries({ queryKey: ['trust-endorsements'] });
       void queryClient.invalidateQueries({ queryKey: ['trust-budget'] });
+      void queryClient.invalidateQueries({ queryKey: ['trust-scores'] });
     },
   });
 }
