+++ web/src/features/trust/api/client.ts
@@ -1,12 +1,29 @@
 /**
  * Trust API client
- * Type-safe REST client for trust scores, budget, endorsements, and invites
+ * Type-safe REST client for trust scores, budget, endorsements, and invites.
+ *
+ * Shared API functions (budget, invites, endorsements) live in @/api/trust and
+ * @/api/endorsements so the endorsements feature can also import them without
+ * violating the no-cross-feature-import ESLint rule.
  */
 
 import { signedFetchJson } from '@/api/signing';
 import type { CryptoModule } from '@/providers/CryptoProvider';
 
-// === Types ===
+// Re-export shared types and functions so feature consumers can import from
+// the trust feature barrel without knowing the internal split.
+export type { Endorsement, EndorsementsListResponse } from '@/api/endorsements';
+export { getMyEndorsements } from '@/api/endorsements';
+export type { TrustBudget, Invite, AcceptInviteResult, CreateInvitePayload } from '@/api/trust';
+export {
+  getMyBudget,
+  createInvite,
+  listMyInvites,
+  acceptInvite,
+  revokeEndorsement,
+} from '@/api/trust';
+
+// === Trust-feature-only types ===
 
 export interface Denouncement {
   id: string;
@@ -33,46 +50,13 @@ export interface ScoreSnapshot {
   computed_at: string;
 }
 
-export interface TrustBudget {
-  slots_total: number;
-  slots_used: number;
-  slots_available: number;
-  denouncements_total: number;
-  denouncements_used: number;
-  denouncements_available: number;
-}
-
-export interface Invite {
-  id: string;
-  delivery_method: string;
-  accepted_by: string | null;
-  expires_at: string;
-  accepted_at: string | null;
-}
-
-export interface AcceptInviteResult {
-  endorser_id: string;
-  accepted_at: string;
-}
-
 export interface EndorsePayload {
   subject_id: string;
   weight: number;
   attestation: Record<string, unknown>;
 }
 
-export interface RevokePayload {
-  subject_id: string;
-}
-
-export interface CreateInvitePayload {
-  envelope: string;
-  delivery_method: string;
-  relationship_depth?: string;
-  attestation: Record<string, unknown>;
-}
-
-// === API functions ===
+// === Trust-feature-only API functions ===
 
 export async function getMyScores(
   deviceKid: string,
@@ -80,46 +64,6 @@ export async function getMyScores(
   wasmCrypto: CryptoModule
 ): Promise<ScoreSnapshot[]> {
   return signedFetchJson('/trust/scores/me', 'GET', deviceKid, privateKey, wasmCrypto);
-}
-
-export async function getMyBudget(
-  deviceKid: string,
-  privateKey: CryptoKey,
-  wasmCrypto: CryptoModule
-): Promise<TrustBudget> {
-  return signedFetchJson('/trust/budget', 'GET', deviceKid, privateKey, wasmCrypto);
-}
-
-export async function createInvite(
-  deviceKid: string,
-  privateKey: CryptoKey,
-  wasmCrypto: CryptoModule,
-  payload: CreateInvitePayload
-): Promise<Invite> {
-  return signedFetchJson('/trust/invites', 'POST', deviceKid, privateKey, wasmCrypto, payload);
-}
-
-export async function listMyInvites(
-  deviceKid: string,
-  privateKey: CryptoKey,
-  wasmCrypto: CryptoModule
-): Promise<Invite[]> {
-  return signedFetchJson('/trust/invites/mine', 'GET', deviceKid, privateKey, wasmCrypto);
-}
-
-export async function acceptInvite(
-  deviceKid: string,
-  privateKey: CryptoKey,
-  wasmCrypto: CryptoModule,
-  inviteId: string
-): Promise<AcceptInviteResult> {
-  return signedFetchJson(
-    `/trust/invites/${inviteId}/accept`,
-    'POST',
-    deviceKid,
-    privateKey,
-    wasmCrypto
-  );
 }
 
 export async function endorse(
@@ -161,14 +105,4 @@ export async function lookupAccount(
     privateKey,
     wasmCrypto
   );
-}
-
-export async function revokeEndorsement(
-  deviceKid: string,
-  privateKey: CryptoKey,
-  wasmCrypto: CryptoModule,
-  subjectId: string
-): Promise<void> {
-  const payload: RevokePayload = { subject_id: subjectId };
-  await signedFetchJson('/trust/revoke', 'POST', deviceKid, privateKey, wasmCrypto, payload);
 }
