+++ web/src/features/verification/api/index.ts
@@ -1,4 +1,4 @@
-export { getMyEndorsements } from './client';
-export type { Endorsement, EndorsementsListResponse } from './client';
+export { getMyEndorsements } from '@/api/endorsements';
+export type { Endorsement, EndorsementsListResponse } from '@/api/endorsements';
 export { useVerificationStatus } from './queries';
 export type { VerificationStatus } from './queries';
